Rename wayland_shell_surface.{c,h} to shell_surface.{c,h}
