+++ src/compositor/wayland/shell.c
@@ -30,7 +30,7 @@
 
 #include "compositor/wayland/surface.h"
 #include "compositor/wayland/shell.h"
-#include "compositor/wayland/wayland_shell_surface.h"
+#include "compositor/wayland/shell_surface.h"
 #include "util/arithmetical.h"
 #include "util/wayland.h"
